pyproject.toml
[tool.poetry]
name = "TransferBot"
version = "0.1.0"
description = ""
authors = ["Vyacheslav Kostrov <slavkostrov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "2.24"
aiohttp = "3.8.3"
aiosignal = "1.3.1"
async-timeout = "4.0.2"
attrs = "22.2.0"
babel = "2.9.1"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
colorama = "0.4.6"
contourpy = "1.0.6"
cycler = "0.11.0"
exceptiongroup = "1.1.0"
fonttools = "4.38.0"
frozenlist = "1.3.3"
idna = "3.4"
iniconfig = "2.0.0"
kiwisolver = "1.4.4"
magic-filter = "1.0.9"
matplotlib = "3.6.2"
multidict = "6.0.4"
numpy = "1.24.1"
packaging = "22.0"
pillow = "9.4.0"
pluggy = "1.0.0"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
pytz = "2022.7"
requests = "2.28.1"
setuptools = "57.0.0"
six = "1.16.0"
tomli = "2.0.1"
torch = "1.13.1"
torchaudio = "0.13.1"
torchvision = "0.14.1"
typing-extensions = "4.4.0"
urllib3 = "1.26.13"
wheel = "0.36.2"
yarl = "1.8.2"
trio = "^0.24.0"
anyio = "^4.2.0"
pytest-cov = "^4.1.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "0.20.3"
pytest = "7.2.1"
flake8 = "^7.0.0"
ruff = "^0.2.1"
pre-commit = "^3.6.1"
aiogram-unittest = "^0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
lint.extend-select = [
  "E501",
  "UP",  # pyupgrade
  "I",   # isort
]
# extend-include = ["*.ipynb"]
extend-exclude = ["tests"]
